Bugfixes in translation routine
There was a possibility of an infinite loop in pib.cli.translate_pib,
which created issues once the Punkt tokenizer was introduced. The
routine is not embedded with additional assertions to ensure that at
least one step of an entry is taken in the loop and that no empty batch
is returned, to ensure no silent failure in other parts.

This could be perhaps what led to issues (wrong offset) and wrong
translations in PIB round 2 during CoDs COMAD, eventually fixed by
adding article information in the pipeline. A stronger queue based
translate mechanism with a buffer in place could be implemented for
better batching and faster translation of PIB articles.